+++ modules/core/src/main/scala/moe/brianhsu/live2d/exception/MocInconsistentException.scala
@@ -1,0 +1,3 @@
+package moe.brianhsu.live2d.exception
+
+class MocInconsistentException(filePath: String) extends Exception(s"The loaded .moc3 $filePath is inconsistent")