+++ CHANGELOG.md
@@ -12,8 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `csmGetParameterTypes` function to `NativeCubismAPI` trait.
+- Added `csmHasMocConsistency' function to 'NativeCubismAPI` trait.
+
 - Added `parameterType` field to Parameter trait, to inidicate is this parameter a normal one or a blend shape one.
 - Added `revivedMoc` and `mocVersion` field to `MocInfo` class.
+- Added 'shouldCheckConsistent' parameter to constructor of `MocInfoFileReader` / `AvatarFileReader`, which controls should we check consistency of the loaded `.moc3` file or not.
 
 ### Changed